[ADD] a new config for better format code
