+++ .eslintrc.js
@@ -1,0 +1,14 @@
+module.exports = {
+	root: true,
+	env: {
+		browser: true,
+		es2021: true,
+	},
+	extends: ['airbnb-base', 'prettier'],
+	overrides: [],
+	parserOptions: {
+		ecmaVersion: 'latest',
+		sourceType: 'module',
+	},
+	rules: {},
+};
