+++ .prettierrc.json
@@ -1,0 +1,6 @@
+{
+	"trailingComma": "es5",
+	"tabWidth": 2,
+	"semi": true,
+	"singleQuote": true
+}
