+++ package.json
@@ -1,29 +1,37 @@
 {
-  "name": "pokedex",
-  "version": "1.0.0",
-  "description": "PokeDex",
-  "main": "tailwind.config.js",
-  "dependencies": {},
-  "devDependencies": {
-    "tailwindcss": "^3.2.4"
-  },
-  "scripts": {
-    "css": "npx tailwindcss -i ./public/styles.css -o ./public/dist/styles.css --watch",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Kevin-Illu/PokeAPi.git"
-  },
-  "keywords": [
-    "JS",
-    "HTML",
-    "Tailwindcss"
-  ],
-  "author": "all",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/Kevin-Illu/PokeAPi/issues"
-  },
-  "homepage": "https://github.com/Kevin-Illu/PokeAPi#readme"
+	"name": "pokedex",
+	"version": "1.0.0",
+	"description": "PokeDex",
+	"main": "tailwind.config.js",
+	"dependencies": {},
+	"devDependencies": {
+		"eslint": "^8.28.0",
+		"eslint-config-airbnb-base": "^15.0.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-plugin-import": "^2.26.0",
+		"husky": "^8.0.2",
+		"prettier": "2.7.1",
+		"tailwindcss": "^3.2.4"
+	},
+	"scripts": {
+		"css": "npx tailwindcss -i ./public/styles.css -o ./public/dist/styles.css --watch",
+		"lint": "npx eslint ./",
+		"format": "npx prettier --write .",
+		"test": "echo \"Error: no test specified\" && exit 1"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/Kevin-Illu/PokeAPi.git"
+	},
+	"keywords": [
+		"JS",
+		"HTML",
+		"Tailwindcss"
+	],
+	"author": "all",
+	"license": "ISC",
+	"bugs": {
+		"url": "https://github.com/Kevin-Illu/PokeAPi/issues"
+	},
+	"homepage": "https://github.com/Kevin-Illu/PokeAPi#readme"
 }
