+++ src/components/Date.js
@@ -1,3 +1,5 @@
-export function showCurrentDate() {
-  const div = document.querySelector("#date");
+export default function showCurrentDate() {
+	const div = document.querySelector('#date');
+	const date = new Date();
+	div.appendChild(date);
 }
