+++ tailwind.config.js
@@ -1,8 +1,8 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ["./public/**/*.{html,js}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
+	content: ['./public/**/*.{html,js}'],
+	theme: {
+		extend: {},
+	},
+	plugins: [],
+};
